refactor: migrate validAuthData and validRegisterData
